% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDllcpp11-package.R
\docType{package}
\name{testDllcpp11-package}
\alias{testDllcpp11}
\alias{testDllcpp11-package}
\title{testDllcpp11: Test package for compiling DLLs that link to cpp11}
\description{
Test package for compiling DLLs that link to cpp11. Just for testing purposes.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda
}

}
